refactor(form): split FormAnchor (P12d)
Decompose 406-LoC FormAnchor god-module into 3 cohesive sub-modules
behind a 15-LoC re-export facade. No public-API change: the four
exports (IFormAnchor type, discoverFormAnchor, scopeCandidate,
scopeCandidates) keep their import paths byte-identical via the
facade.

## Why

Phase 12d cluster: FormAnchor.ts was 406 lines mixing two concerns:
ancestor-walking form discovery and SelectorCandidate scoping. KG-
analysis flagged it as a HIGH-tier complex file. Splitting clarifies
the two domains and keeps each file under the CLEAN_CODE.md cap.

## What

- Anchor/AnchorTypes.ts: shared types (IFormAnchor, IAncestorMeta,
  AncestorTuple) + constants (EMPTY_RESULT, FORM_ID_RE,
  NON_FILLABLE_FILTER, ANCESTOR_XPATH, MIN_FILLABLE_INPUTS).
- Anchor/AnchorWalk.ts: DOM ancestor walk + discoverFormAnchor
  (tupleToMeta, mapAncestorTuples, collectAncestorMeta,
  buildSelectorFromMeta, evaluateFormWalk).
- Anchor/AnchorScope.ts: SelectorCandidate rewrites + scopeCandidate
  + scopeCandidates (scopeCss/Placeholder/AriaLabel/Name,
  scopeXpath, three buildScoped*Xpath builders, scopeTextCandidate,
  scopeCssCandidate, SCOPE_BUILDERS, TEXT_SCOPE_BUILDERS,
  tryExtractFormId).
- FormAnchor.ts: 15-LoC facade re-exporting the public surface.

## Guideline compliance

- [x] coding-principle: SRP — each sub-module owns one concern.
- [x] eslint-rules §3: no eslint-disable used.
- [x] eslint-rules §5: refactor commit does NOT touch eslint.config.
- [x] before-commit: never bypass hooks, never `git add .`, atomic.
- [x] commit §5: subject 39 chars; body wraps at <=72.
- [x] comments-in-code: BLUF JSDoc preserved on every export.
- [x] j-doc: @param/@return order; <p> for body paragraphs.

## Tracker(s)

- Phase 12d (Login + Form cluster decoupling — file 7 of 8)
- PVTI_lAHOAPtF084BZbzwzgvauEI (GH Project #3)
- Master plan row 12 — v8.5.0 wave 2
